Moved compact_log_test tests to manifest_test
